Move MarkDistinctNode to SPI
