+++ presto-main/src/main/java/com/facebook/presto/sql/planner/ExpressionDeterminismEvaluator.java
@@ -24,6 +24,7 @@ import static java.util.Objects.requireNonNull;
 
 /**
  * Determines whether a given Expression is deterministic
+ *
  * @deprecated Use {@link com.facebook.presto.sql.relational.RowExpressionDeterminismEvaluator}
  */
 public final class ExpressionDeterminismEvaluator
