+++ presto-main/src/main/java/com/facebook/presto/sql/planner/plan/InternalPlanVisitor.java
@@ -114,11 +114,6 @@ public abstract class InternalPlanVisitor<R, C>
         return visitPlan(node, context);
     }
 
-    public R visitMarkDistinct(MarkDistinctNode node, C context)
-    {
-        return visitPlan(node, context);
-    }
-
     public R visitGroupId(GroupIdNode node, C context)
     {
         return visitPlan(node, context);
