+++ presto-spi/src/main/java/com/facebook/presto/spi/plan/PlanVisitor.java
@@ -65,6 +65,11 @@ public abstract class PlanVisitor<R, C>
         return visitPlan(node, context);
     }
 
+    public R visitMarkDistinct(MarkDistinctNode node, C context)
+    {
+        return visitPlan(node, context);
+    }
+
     public R visitTopN(TopNNode node, C context)
     {
         return visitPlan(node, context);
